% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msign.R
\name{mod.sign.test}
\alias{mod.sign.test}
\title{Modified sign test when the center of symmetry is unknown}
\usage{
mod.sign.test(x)
}
\arguments{
\item{x}{data set need to be tested}
}
\value{

}
\description{
Testing whether a data set comes from a symmetric distribution when the center of symmetry is unknown.
}
\examples{
x = rchisq(100,1)
mod.sign.test(x)


}
