% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsym.R
\name{mod.sym.test}
\alias{mod.sym.test}
\title{Modified symmetric test}
\usage{
mod.sym.test(
  x,
  y = NULL,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater"),
  method = "W"
)
}
\arguments{
\item{x}{data set to be tested}

\item{y}{another data set}

\item{paired}{logical. whether data set x and y are paired}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two sided" (default), "greater", or "less".
greater: test whether positively skewed (right-skewed),
less : test whether negatively skewed (left-skewed)}

\item{method}{a character string specifying which symmetric test to be used, "W" is modified wilcoxon sign rank test,
and "S" is modified sign test.}
}
\value{

}
\description{
Provides one symmetric test that incorporated 2 methods, "W" and "S". Also, paired data can be tested.
}
\examples{
x = rchisq(100,1)
y = rchisq(100,1)
mod.sym.test(x, method="W")
mod.sym.test(x, method="S", alternative="greater")

mod.sym.test(x, y, paired=T, method="W")

}
