% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwilcox.R
\name{mod.wilcox.test}
\alias{mod.wilcox.test}
\title{Modified Wilcoxon test when the center of symmetry is unknown.}
\usage{
mod.wilcox.test(x)
}
\arguments{
\item{x}{data set need to be tested}
}
\value{

}
\description{
Testing whether a data set comes from a symmetric distribution when the center of symmetry is unknown.
}
\examples{
x = rchisq(100,1)
mod.wilcox.test(x)



}
